*hannes: long vs. int
git-svn-id: file:///usr/local/Singular/svn/trunk@12119 2c84dea3-7e68-4137-9b89-c4e89433aadc

+++ kernel/modulop.cc
@@ -1,7 +1,7 @@
 /****************************************
 *  Computer Algebra System SINGULAR     *
 ****************************************/
-/* $Id: modulop.cc,v 1.17 2009-09-16 12:26:26 Singular Exp $ */
+/* $Id: modulop.cc,v 1.18 2009-09-21 13:59:32 Singular Exp $ */
 /*
 * ABSTRACT: numbers modulo p (<=32003)
 */
@@ -74,8 +74,8 @@ number npInit (int i)
 */
 int npInt(number &n, const ring r)
 {
-  if ((long)n > (r->ch >>1)) return (int)((long)n -(r->ch));
-  else                       return (int)((long)n);
+  if ((long)n > (((long)r->ch) >>1)) return (int)((long)n -((long)r->ch));
+  else                               return (int)((long)n);
 }
 
 number npAdd (number a, number b)
